Refactoring current service and controllers
